I18n for strings, that are read from the configuration file. L13n of jabberd14 to German. Installing German language file on make install.
automake 1.5+ is no required (at least 1.4 does not work anymore). You only need this to build jabberd14 from SVN. If you want to install jabberd14 on a server, that does only have automake 1.4, you can "make dist" the package on a system having newer automake and install the resulting tar-ball on the target system. I am sorry for this change, but the compatibility fix for automake 1.4 I introduced some days ago was not satisfying.