Working on GnuTLS support, but nothing really advanced from the last commit yet. GnuTLS currently is not able to look-up id-on-xmppAddr values in a certificate (we need them for Jabber). So I am waiting for a solution in GnuTLS and have contacted the author of this library.
Updated the pool debugging functions, to output additional data: the number of pools, the total amount of memory allocated in pools, the size of the biggest pool. What I want to know is if the growing memory consumption in jabberd14 is caused by pools not being disallocated (in that case the number of pools should not stop growing), or if there are pools, that allocate more and more memory (in which case there are long-living pools, that are used for data, that is probably not needed for a long time, but only for processing a stanza or a session), or if the total memory in pools differs very much from what top reports, we would have a real memory leak, meaning that memory is allocated outside of pools, that is not freed (maybe by a library we are using, or inside the few cases where we allocate memory directly).

Compatibility improvements with old compilers (pre ANSI C99).