Updated TODO items for the time after 1.6.1 has been released.
Increased importance for removing the dependancy of libpth, most likely changing to pthreads.
Added TODO itmes for new APIs to write server components and session manager modules, one to provide a C++ OOP API, the other one to provide an API to a scripting language. The scripting language is not yet selected. I know how to integrate the Perl or the Python interpreter. Ruby seems to be interesting as well, but I am not that familar with that language yet.

BTW: The last revision (1468) is running on amessage sucessfully again. No need to stick with revision 1460 anymore.