New feature: Administrator of the server can configure xpath expressions, that are used against routed stanzas in the server. If the expression matches, the stanza will be logged. This feature is thought to be used for debugging purposes.
Modified behaviour:
If a packet from a client connection manager to a session manager using the session control protocol bounces, the bounce will now trigger the client connection manager to close the connection to the client. This matches the behaviour with the old way these two components communicate and is reasonable, as bounced packages normally mean, that the packet is for a session, that does not exist (anymore).