Fixing two problems:
1. (compile problem)

jabberd/mio_tls.cc had been adapted to newer version of GnuTLS where function signatures for some OpenPGP functions had been changed. This breaks compiling with older versions of GnuTLS. Therefore we now check if the new function signatures are available. If not, we do disable some OpenPGP functionality in jabberd14.

2. (runtime problem)

When using mod_dynamic.cc the module tries to load the dynamically added domains on startup. This had problems when the the xdb module had not yet been started when the JSM starts. mod_dynamic.cc now retries the xdb request, it if fails at the first time. This is a intermediate fix. I will try to get some logic for the order of starting modules, that will fix this problem from the ground.