Getting jabberd compiled again.
This commit makes the compile work on a current amd64 linux system.
However jabberd is not running again.

There is a problem that that symbols in the ELF tables don't completely
match. I still have to investigate into this.

Also there are a lot of compilter warning that have to do with changed
lengths of some times on a amd64 system, that I still have to fix.